Beginning to format and re-working editing flow
I can't take all this ugly unformatted text any longer, so I'm starting to make it look nice.
Also, the workflow for creating and editing a character had some problems. The new method allows character creation to
be undertaken in alternate orders, if desired, and will make editing an already-created (or half-created) character simpler
to implement.